refactor: merge favorites and timeline pagination
