--- domain/content/pagination/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/content/pagination/FavoritesPaginationSpecification.kt
@@ -1,11 +1,0 @@
-package com.livefast.eattrash.raccoonforfriendica.domain.content.pagination
-
-sealed interface FavoritesPaginationSpecification {
-    data class Favorites(
-        val includeNsfw: Boolean = true,
-    ) : FavoritesPaginationSpecification
-
-    data class Bookmarks(
-        val includeNsfw: Boolean = true,
-    ) : FavoritesPaginationSpecification
-}
