+++ domain/content/pagination/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/content/pagination/TimelinePaginationSpecification.kt
@@ -30,4 +30,12 @@ sealed interface TimelinePaginationSpecification {
         val userId: String,
         val includeNsfw: Boolean = true,
     ) : TimelinePaginationSpecification
+
+    data class Favorites(
+        val includeNsfw: Boolean = true,
+    ) : TimelinePaginationSpecification
+
+    data class Bookmarks(
+        val includeNsfw: Boolean = true,
+    ) : TimelinePaginationSpecification
 }
